social_story_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
langchain==0.2.14
langgraph==0.2.24
openai==1.48.0
